+++ progfiguration/newsite/exgroup.py
@@ -1,0 +1,11 @@
+"""Example group"""
+
+from progfiguration.progfigtypes import Bunch
+
+group = Bunch(
+    roles=Bunch(
+        settz={
+            "timezone": "UTC",
+        }
+    ),
+)